--- a/buffer/src/main/java/io/netty/buffer/AbstractDerivedByteBuf.java
+++ b/buffer/src/main/java/io/netty/buffer/AbstractDerivedByteBuf.java
@@ -32,6 +32,11 @@ public abstract class AbstractDerivedByteBuf extends AbstractByteBuf {
     }
 
     @Override
+    final boolean isAccessible() {
+        return unwrap().isAccessible();
+    }
+
+    @Override
     public final int refCnt() {
         return refCnt0();
     }