--- a/buffer/src/main/java/io/netty/buffer/CompositeByteBuf.java
+++ b/buffer/src/main/java/io/netty/buffer/CompositeByteBuf.java
@@ -2123,6 +2123,11 @@ public class CompositeByteBuf extends AbstractReferenceCountedByteBuf implements
     }
 
     @Override
+    boolean isAccessible() {
+        return componentCount != -1;
+    }
+
+    @Override
     public ByteBuf unwrap() {
         return null;
     }