--- a/buffer/src/main/java/io/netty/buffer/SwappedByteBuf.java
+++ b/buffer/src/main/java/io/netty/buffer/SwappedByteBuf.java
@@ -998,6 +998,11 @@ public class SwappedByteBuf extends ByteBuf {
     }
 
     @Override
+    final boolean isAccessible() {
+        return buf.isAccessible();
+    }
+
+    @Override
     public ByteBuf retain() {
         buf.retain();
         return this;