--- a/buffer/src/main/java/io/netty/buffer/WrappedByteBuf.java
+++ b/buffer/src/main/java/io/netty/buffer/WrappedByteBuf.java
@@ -1035,7 +1035,7 @@ class WrappedByteBuf extends ByteBuf {
     }
 
     @Override
-    boolean isAccessible() {
+    final boolean isAccessible() {
         return buf.isAccessible();
     }
 }