Motivation:

Some profiles of grpc-java indicate that overhead from `AbstractByteBuf.checkAccessible()` is quite significant due the relatively fine-grained way that `ByteBuf`s are accessed by the http2 hpack logic (5-10% of an entire end-to-end roundtrip benchmark!)

This PR takes the great improvement done in #8266 a bit further, simpliyfing the non-volatile/best-effort check. I did not dig deep enough to see why the existing `internalRefCnt()` logic was relatively slower
given that it does appear to get inlined. One observation though is that the prior optimization wasn't applied to derived buffers (slices etc.) which presumably was an oversight - this would be fairly simple to fix without the rest of this change (by overriding/delegating `internalRefCnt`) but that wouldn't help with the non-derived case.

Modifications:

- Replace `AbstractByteBuf.internalRefCnt()` with an internal (package-private) `ByteBuf.isAccessible()` method; the default impl still checks for nonzero `refCnt()`
- Set `maxCapacity` to -1 upon deallocation in `AbstractRefCountedByteBuf`, and add a final override of `isAccessible()` which checks this
- Appropriately forward `isAccessible()` invocations in `AbstractDerivedByteBuf`/`WrappedByteBuf`/`SwappedByteBuf` (also final overrides)
- Have `CompositeByteBuf` set `numComponents` to -1 upon deallocation so that the existing component index checks done on most methods also serve to check accessibility
- Add `ByteBufAccessBenchmark` which is an extension of `UnsafeByteBufBenchmark` (maybe latter could now be removed)

This change is similar to @ejona86's suggestion in #8266 (making use of `maxCapacity`), and I'd also thought of "invalidating" the reader/writer indices upon deallocation and having existing bounds checks do double-duty where possible. That could be a worthwhile addition irrespective of whether `checkAccessible` is otherwise changed to be false by default, but decided to keep this iteration minimally invasive. A complication is that bounds checking logic is in `AbstractByteBuf` but the index-invalidation-upon-deallocate logic must live in `AbstractReferenceCountedByteBuf` meaning that some further rework would be needed to avoid breaking the checks for any third-party impls that extend `AbstractByteBuf` directly. It's also the case that not all of the bounds checks involve these fields (absolute ones currently only use `capacity()` which is impl-specific),

Results:

Non-negligible speedups, especially for `CompositeByteBuf` which is twice as fast in some cases.

Before:
```
Benchmark                              (bufferType)   Mode  Cnt
Score         Error  Units
ByteBufAccessBenchmark.setGetLong            UNSAFE  thrpt   30 242844273.620 ± 1778416.963  ops/s
ByteBufAccessBenchmark.setGetLong      UNSAFE_SLICE  thrpt   30 184970819.249 ± 1990964.278  ops/s
ByteBufAccessBenchmark.setGetLong              HEAP  thrpt   30 242099517.501 ± 2661438.140  ops/s
ByteBufAccessBenchmark.setGetLong         COMPOSITE  thrpt   30  76961466.906 ± 1318414.178  ops/s
ByteBufAccessBenchmark.setGetLong               NIO  thrpt   30 279181674.131 ± 2162611.704  ops/s
```

After:
```
Benchmark                              (bufferType)   Mode  Cnt
Score          Error  Units
ByteBufAccessBenchmark.setGetLong            UNSAFE  thrpt   30 285754526.496 ±  3292540.505  ops/s
ByteBufAccessBenchmark.setGetLong      UNSAFE_SLICE  thrpt   30 209794441.861 ±  1646039.479  ops/s
ByteBufAccessBenchmark.setGetLong              HEAP  thrpt   30 283203453.736 ±  1987854.916  ops/s
ByteBufAccessBenchmark.setGetLong         COMPOSITE  thrpt   30 155437066.991 ±  1492219.965  ops/s
ByteBufAccessBenchmark.setGetLong               NIO  thrpt   30 277362327.090 ±  3070532.642  ops/s
```