@njhill 2019-02-22
The knock-on improvement to HPACK benchmarks is also evident:

Before:
```
Benchmark                     (limitToAscii)  (sensitive)  (size)   Mode  Cnt        Score       Error  Units
HpackDecoderBenchmark.decode            true         true   SMALL  thrpt   30   370410.796 ±  4110.021  ops/s
HpackDecoderBenchmark.decode            true         true  MEDIUM  thrpt   30    49068.535 ±   406.485  ops/s
HpackDecoderBenchmark.decode            true        false   SMALL  thrpt   30   350153.682 ±  5251.114  ops/s
HpackDecoderBenchmark.decode            true        false  MEDIUM  thrpt   30    48853.463 ±   496.576  ops/s
HpackDecoderBenchmark.decode           false         true   SMALL  thrpt   30  1731930.970 ± 13894.478  ops/s
HpackDecoderBenchmark.decode           false         true  MEDIUM  thrpt   30   493797.242 ±  6731.768  ops/s
HpackDecoderBenchmark.decode           false        false   SMALL  thrpt   30  1573771.081 ± 13874.584  ops/s
HpackDecoderBenchmark.decode           false        false  MEDIUM  thrpt   30   428455.157 ±  6313.333  ops/s
```

After:
```
Benchmark                     (limitToAscii)  (sensitive)  (size)   Mode  Cnt        Score       Error  Units
HpackDecoderBenchmark.decode            true         true   SMALL  thrpt   30   378633.144 ± 10188.415  ops/s
HpackDecoderBenchmark.decode            true         true  MEDIUM  thrpt   30    58525.564 ±   329.438  ops/s
HpackDecoderBenchmark.decode            true        false   SMALL  thrpt   30   371989.033 ±  3195.292  ops/s
HpackDecoderBenchmark.decode            true        false  MEDIUM  thrpt   30    59141.990 ±   438.444  ops/s
HpackDecoderBenchmark.decode           false         true   SMALL  thrpt   30  1963143.732 ± 16181.923  ops/s
HpackDecoderBenchmark.decode           false         true  MEDIUM  thrpt   30   571016.787 ±  5614.790  ops/s
HpackDecoderBenchmark.decode           false        false   SMALL  thrpt   30  1891541.810 ± 19540.314  ops/s
HpackDecoderBenchmark.decode           false        false  MEDIUM  thrpt   30   533891.353 ±  4803.908  ops/s
```