@normanmaurer 2019-02-22
@njhill imho this change can not go into 4.1 as it would change semantics of `maxCapacity()` to not return the correct value anymore when deallocated.