@normanmaurer 2019-02-22
@njhill as stated by you we should try to understand why this change is faster then what we had before as it should be the same speed imho... That said please provide a PR to fix the derived buffers :)