@njhill 2019-02-22
@normanmaurer I thought that `maxCapacity()` would not really be meaningful for a deallocated buffer (in fact I thought all bets are off generally in this situation). For example, `PooledByteBuf`s reset this and the indicies when they are reused.

However I also can think of a trick to work around this, will push an update to see what you think. Worst case like you say I think there's some (possibly smaller) improvement here possible without relying on that field.