@normanmaurer 2019-02-22
@njhill that's a good point that said I am still a bit afraid of the change here. I guess what you could do here is to make the value negative on release and then use `abs` when the user queries it. That said I still would love to understand why this is even needed as it theory the old behaviour should be good enough.