@njhill 2019-02-22
@normanmaurer yep I already made the `abs` change, just want to benchmark before pushing it here. But I will open another PR regardless with related improvements that don't involve use of a different field. When I briefly tried this out before it appeared to give roughly half the gains.