@normanmaurer 2019-02-23
@njhill any changes in the benchmark results after the abs change ? Also before pulling this in I would like to understand why this is needed and the current approach is slower. In theory I think it should not