@njhill 2019-02-23
@normanmaurer benchmark with the `Math.abs` commit shows very little difference (possibly slightly slower). I agree this PR should still be considered tentative pending investigation/understanding of why it's necessary.

I've created a separate branch with some similar improvements to this one but avoiding use of `maxCapacity` - see cd26d4f394978da0ba58e04e3932ebb9e693cfc9. First benchmarks are similar to what I observed before: the gains are much smaller than those shown above for this PR. Some deeper analysis is needed I guess...