@franz1981 2019-02-23
I've some numbers to share, 4.1:
```
Benchmark                              (bufferType)  (checkAccessible)  (checkBounds)   Mode  Cnt          Score         Error  Units
ByteBufAccessBenchmark.readEightBytes     COMPOSITE               true           true  thrpt   20   14655020.539 ±  425320.817  ops/s
ByteBufAccessBenchmark.readEightBytes     COMPOSITE               true          false  thrpt   20   17807254.275 ±  193070.149  ops/s
ByteBufAccessBenchmark.readEightBytes     COMPOSITE              false           true  thrpt   20   24776669.955 ±  269819.385  ops/s
ByteBufAccessBenchmark.readEightBytes     COMPOSITE              false          false  thrpt   20   28118611.519 ±  938274.799  ops/s
ByteBufAccessBenchmark.setGetLong         COMPOSITE               true           true  thrpt   20   72695175.133 ± 1029659.304  ops/s
ByteBufAccessBenchmark.setGetLong         COMPOSITE               true          false  thrpt   20   85866095.841 ±  955278.291  ops/s
ByteBufAccessBenchmark.setGetLong         COMPOSITE              false           true  thrpt   20   95962385.554 ±  356967.006  ops/s
ByteBufAccessBenchmark.setGetLong         COMPOSITE              false          false  thrpt   20  118367539.734 ± 7501704.810  ops/s
ByteBufAccessBenchmark.setLong            COMPOSITE               true           true  thrpt   20   74363920.616 ± 6862737.577  ops/s
ByteBufAccessBenchmark.setLong            COMPOSITE               true          false  thrpt   20   76531968.295 ± 1184693.315  ops/s
ByteBufAccessBenchmark.setLong            COMPOSITE              false           true  thrpt   20   80340118.090 ±  321379.275  ops/s
ByteBufAccessBenchmark.setLong            COMPOSITE              false          false  thrpt   20   90684143.719 ±  935820.121  ops/s
```