@njhill 2019-02-23
@franz1981 I like the of including the checks-disabled cases in the benchmark as a point of comparison, but for now maybe it simplifies things to exclude them.

What's of most interest imho is improvement we can get in the default setting case, and in particular understanding why checking a separate plain field is faster than the existing approach... i.e. some perfasm comparison between this PR and the other branch referenced above (which narrows the deltas).

I will have a go at this when I get a chance (have some other weekend commitments), but feel free to do so too in the meantime! Thanks again!