@franz1981 2019-02-23
@njhill 
I've a strong feeling it is dependent by something related to CHA of the JIT instead: if you change `AbstractByteBuf` like this:
```
    protected int nonVolatileRawCnt() {
        throw new UnsupportedOperationException();
    }

    /**
     * Should be called by every method that tries to access the buffers content to check
     * if the buffer was released before.
     */
    protected final void ensureAccessible() {
        if (checkAccessible && AbstractReferenceCountedByteBuf.realRefCnt(nonVolatileRawCnt()) == 0) {
            throw new IllegalReferenceCountException(0);
        }
    }
```
making `AbstractReferenceCountedByteBuf.realRefCnt` public, you will see that the performance are similar to this PR. I hope soon to get some ASM to check :+1: