@njhill 2019-02-24
@normanmaurer class hierarchy analysis ... I had to look it up too :)