@franz1981 2019-02-24
@normanmaurer Nick has preceded me re CHA :P
I've tried to get an ASM log to be used with JitWatch with no success...seems that I cannot print assembly on the JMH bench...